separate constants to different files
